requirements.txt
google-cloud-storage
matplotlib
numpy
opencv-python
pyav
python-dotenv
python-magic